nem: mosaics moved to seperate file
